refactor(Price): 정적팩토리메서드 파라미터 타입 변경(BigDecimal -> long)

+++ src/main/java/kitchenpos/menu/application/MenuService.java
@@ -36,7 +36,7 @@ public class MenuService {
 
     @Transactional
     public MenuResponse create(final MenuRequest menuRequest) {
-        Price price = Price.from(BigDecimal.valueOf(menuRequest.getPrice()));
+        Price price = Price.from(menuRequest.getPrice());
         MenuGroup menuGroup = findMenuGroup(menuRequest);
         List<MenuProduct> menuProducts = createMenuProducts(menuRequest.getMenuProducts());
         Menu menu = Menu.createMenu(menuRequest.getName(), price, menuGroup, menuProducts);
