+++ src/test/java/kitchenpos/menu/domain/MenuProductsTest.java
@@ -13,8 +13,8 @@ public class MenuProductsTest {
     @Test
     @DisplayName("메뉴 상품의 가격 총합을 구한다.")
     void calculateTotalPrice() {
-        Product product = Product.of("허니콤보", BigDecimal.valueOf(19_000L));
-        Product otherProduct = Product.of("레드콤보", BigDecimal.valueOf(20_000L));
+        Product product = Product.of("허니콤보", 19_000L);
+        Product otherProduct = Product.of("레드콤보", 20_000L);
         MenuProduct menuProduct = MenuProduct.createMenuProduct(product, 1L);
         MenuProduct otherMenuProduct = MenuProduct.createMenuProduct(otherProduct, 1L);
         MenuProducts menuProducts = new MenuProducts(Lists.list(menuProduct, otherMenuProduct));
