+++ src/test/java/kitchenpos/product/application/ProductServiceTest.java
@@ -11,7 +11,6 @@ import kitchenpos.product.infrastructure.ProductRepository;
 import org.assertj.core.util.Lists;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import java.math.BigDecimal;
 import java.util.List;
 import java.util.stream.Collectors;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -30,7 +29,7 @@ public class ProductServiceTest {
     @Test
     @DisplayName("상품 생성에 성공한다.")
     void create() {
-        Product product = Product.of("허니콤보", BigDecimal.valueOf(19_000L));
+        Product product = Product.of("허니콤보", 19_000L);
         when(productRepository.save(any())).thenReturn(product);
         ProductRequest productRequest = new ProductRequest("허니콤보", 19_000L);
 
@@ -42,8 +41,8 @@ public class ProductServiceTest {
     @Test
     @DisplayName("상품 목록을 조회한다.")
     void findAll() {
-        Product honeyCombo = Product.of("허니콤보", BigDecimal.valueOf(19_000L));
-        Product redCombo = Product.of("레드콤보", BigDecimal.valueOf(19_000L));
+        Product honeyCombo = Product.of("허니콤보", 19_000L);
+        Product redCombo = Product.of("레드콤보", 19_000L);
         when(productRepository.findAll()).thenReturn(Lists.list(honeyCombo, redCombo));
 
         List<ProductResponse> actual = productService.list();
