+++ src/test/java/kitchenpos/product/domain/PriceTest.java
@@ -14,7 +14,7 @@ public class PriceTest {
     @ParameterizedTest(name = "{0} 값으로 가격 객체 생성한다.")
     @ValueSource(longs = {1L, 1000L, 10_000L})
     void create(long value) {
-        Price price = Price.from(BigDecimal.valueOf(value));
+        Price price = Price.from(value);
 
         assertThat(price).isNotNull();
     }
@@ -24,7 +24,7 @@ public class PriceTest {
     @ValueSource(longs = {-1L, -1000L, -10_000L})
     void createFail(long value) {
         assertThatThrownBy(() -> {
-            Price price = Price.from(BigDecimal.valueOf(value));
+            Price price = Price.from(value);
         }).isInstanceOf(IllegalArgumentException.class);
     }
 
@@ -32,7 +32,7 @@ public class PriceTest {
     @ParameterizedTest(name = "{0}과 {1}의 곱은 {2}이다.")
     @CsvSource(value = {"1000:3:3000", "21000:2:42000"}, delimiter = ':')
     void multiply(long value, long multiplyValue, BigDecimal expected) {
-        Price price = Price.from(BigDecimal.valueOf(value));
+        Price price = Price.from(value);
 
         price.multiply(multiplyValue);
 
@@ -43,7 +43,7 @@ public class PriceTest {
     @ParameterizedTest(name = "{0}과 {1}의 곱은 0 미만이다.")
     @CsvSource(value = {"1000:-3", "21000:-2"}, delimiter = ':')
     void multiplyFail(long value, long multiplyValue) {
-        Price price = Price.from(BigDecimal.valueOf(value));
+        Price price = Price.from(value);
 
         assertThatThrownBy(() -> {
             price.multiply(multiplyValue);
